requirements.txt
streamlit
transformers
sentence-transformers
pdfplumber
PyPDF2
scikit-learn
numpy
fpdf 
python-docx
